fix(APPC-4050) vk onboarding navigation; scroll buttons
