+++ app/src/main/res/layout/fragment_onboarding.xml
@@ -46,7 +46,7 @@
           app:layout_constraintEnd_toEndOf="parent"
           app:layout_constraintStart_toStartOf="parent"
           app:layout_constraintTop_toTopOf="parent"
-          app:tint="@color/styleguide_white"
+          app:tint="@color/styleguide_pink"
           />
 
       <include
