+++ app/src/main/res/layout/onboarding_vk_payment_layout.xml
@@ -35,36 +35,47 @@
       tools:visibility="visible"
       />
 
-  <FrameLayout
-      android:id="@+id/complete_payment_view"
+  <ScrollView
       android:layout_width="match_parent"
-      android:layout_height="wrap_content"
-      android:layout_marginBottom="32dp"
-      android:background="@color/styleguide_blue_secondary"
-      android:visibility="gone"
-      app:layout_constraintBottom_toBottomOf="parent"
-      app:layout_constraintEnd_toEndOf="parent"
-      app:layout_constraintStart_toStartOf="parent"
-      app:layout_constraintTop_toTopOf="parent"
-      app:layout_constraintVertical_bias="0.36"
-      >
-    <include
-        android:id="@+id/fragment_first_iab_transaction_completed"
-        layout="@layout/fragment_first_iab_transaction_completed"
-        />
-  </FrameLayout>
+      android:layout_height="match_parent">
 
-  <include
-      android:id="@+id/onboarding_success_vk_buttons"
-      layout="@layout/onboarding_payment_success_buttons"
-      android:layout_width="match_parent"
-      android:layout_height="wrap_content"
-      android:visibility="gone"
-      tools:visibility="visible"
-      app:layout_constraintEnd_toEndOf="parent"
-      app:layout_constraintStart_toStartOf="parent"
-      app:layout_constraintBottom_toBottomOf="parent"
-      />
+    <androidx.constraintlayout.widget.ConstraintLayout
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content">
+
+      <FrameLayout
+          android:id="@+id/complete_payment_view"
+          android:layout_width="match_parent"
+          android:layout_height="wrap_content"
+          android:layout_marginBottom="32dp"
+          android:background="@color/styleguide_blue_secondary"
+          android:visibility="gone"
+          app:layout_constraintBottom_toTopOf="@+id/onboarding_success_vk_buttons"
+          app:layout_constraintEnd_toEndOf="parent"
+          app:layout_constraintStart_toStartOf="parent"
+          app:layout_constraintTop_toTopOf="parent"
+          app:layout_constraintVertical_bias="0.36"
+          >
+        <include
+            android:id="@+id/fragment_first_iab_transaction_completed"
+            layout="@layout/fragment_first_iab_transaction_completed"
+            />
+      </FrameLayout>
+
+      <include
+          android:id="@+id/onboarding_success_vk_buttons"
+          layout="@layout/onboarding_payment_success_buttons"
+          android:layout_width="match_parent"
+          android:layout_height="wrap_content"
+          android:visibility="gone"
+          tools:visibility="visible"
+          app:layout_constraintEnd_toEndOf="parent"
+          app:layout_constraintStart_toStartOf="parent"
+          app:layout_constraintBottom_toBottomOf="parent"
+          />
+
+    </androidx.constraintlayout.widget.ConstraintLayout>
+  </ScrollView>
 
   <include
       android:id="@+id/no_network"
